requirement.txt
Django==1.10.1
djangorestframework==3.4.7
elasticsearch==2.4.0
elasticsearch-dsl==2.1.0
python-dateutil==2.5.3
six==1.10.0
urllib3==1.17
